Add pw check to item delete

+++ src/controllers/ItemController.ts
@@ -130,15 +130,34 @@ const ItemController = (() => {
     res.render('items/delete', { title: 'Delete Item', item: item });
   };
 
-  const destroy_post = async (req: Request, res: Response, next: NextFunction) => {
-    try {
-      await cloudinary.v2.uploader.destroy(req.body.img_id);
-      await Item.findByIdAndDelete(req.body.id).exec();
-      res.redirect('/items');
-    } catch(err) {
-      next(err);
-    }
-  };
+  const destroy_post = [
+    body('admin_pw', 'Admin Password Required')
+      .escape()
+      .trim()
+      .custom((value, { req }) => {
+        if (value === EnvVars.AdminPassword) {
+          return true;
+        } else {
+          return false;
+        }
+      }).withMessage('Wrong Admin Password'),
+    async (req: Request, res: Response, next: NextFunction) => {
+
+      const errors = validationResult(req);
+      if (!errors.isEmpty()) {
+        const item = await Item.findById(req.params.id).exec();
+
+        res.render('items/delete', { title: 'Delete Item', item: item, errors: errors.array()});
+      } else {
+        try {
+          await cloudinary.v2.uploader.destroy(req.body.img_id);
+          await Item.findByIdAndDelete(req.body.id).exec();
+          res.redirect('/items');
+        } catch(err) {
+          next(err);
+        }
+      }
+  }];
 
   const store = [
     upload.single('img'),
